requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-google-genai
google-generativeai
faiss-cpu
PyPDF2
pdf2image
Pillow